Create Question3.cpp
Write a program to find the count of 3 multiples in a given input integer array. 
Include a function named divisibleBy3 that accepts 2 arguments and returns an int. 
The first argument is the input array and the second argument is an int that corresponds to the size of 
the array. The function returns an int that corresponds to the count of 3 multiples. 
 
If the size of the array is negative or if any element in the array is negative, print “Invalid Input”
and terminate the program. 
 
Input and Output Format: 

Input consists of n+1 integers. The first integer corresponds to n, the number of elements in the array. The next 'n' integers correspond to the elements in the array. 
Output consists of an integer that corresponds to the count of 3 multiples